Crate remote_settings_client

Remote Settings Client

A library for fetching and synchronizing Remote Settings data.

Example

  use remote_settings_client::{Client, RingVerifier};
  use viaduct::set_backend;
  use viaduct_reqwest::ReqwestBackend;

  set_backend(&ReqwestBackend).unwrap();

  let mut client = Client::builder()
    .bucket_name("main-preview")
    .collection_name("search-config")
    .verifier(Box::new(RingVerifier {}))
    .build()
    .unwrap();

  client.sync(None).unwrap();

  match client.get() {
    Ok(records) => println!("{:?}", records),
    Err(error) => println!("Error fetching/verifying records: {:?}", error),
  };

See Client for more infos.

Re-exports

pub use client::Client;

pub use client::Collection;

pub use client::Record;

pub use client::DEFAULT_BUCKET_NAME;

pub use client::DEFAULT_SERVER_URL;

Modules

client

Enums

SignatureError

StorageError

Traits

Storage
A trait for giving a type a custom storage implementation

Verification
A trait for signature verification of collection data.