Trait remote_settings_client::Verification

pub trait Verification: Send {
    fn verify(&self, collection: &Collection) -> Result<(), SignatureError>;

    fn fetch_certificate_chain(
        &self, 
        collection: &Collection
    ) -> Result<Vec<u8>, SignatureError> { ... }

    fn decode_signature(
        &self, 
        collection: &Collection
    ) -> Result<Vec<u8>, SignatureError> { ... }

    fn serialize_data(
        &self, 
        collection: &Collection
    ) -> Result<Vec<u8>, SignatureError> { ... }
}

A trait for signature verification of collection data.

You may want to use your own verification implementation (eg. using OpenSSL instead of ring or rc_crypto).

How can I implement Verification?

struct SignatureVerifier {}

impl Verification for SignatureVerifier {
    fn verify(&self, collection: &Collection) -> Result<(), SignatureError> {
        Ok(())
    }
}

let client = Client::builder()
   .collection_name("cid")
   .verifier(Box::new(SignatureVerifier {}))
   .build();

Required methods

fn verify(&self, collection: &Collection) -> Result<(), SignatureError>

Verifies signature for a given Collection struct

Errors

If all the steps are performed without errors, but the specified collection data does not match its signature, then a SignatureError::MismatchError is returned.

If errors occur during certificate download, parsing, or data serialization, then the corresponding error is returned.

Provided methods

fn fetch_certificate_chain(
&self,
collection: &Collection
) -> Result<Vec<u8>, SignatureError>

fn decode_signature(
&self,
collection: &Collection
) -> Result<Vec<u8>, SignatureError>

fn serialize_data(
&self,
collection: &Collection
) -> Result<Vec<u8>, SignatureError>